DAQ 模組大幅重構：增量統計、串流檔案輸出、DaqmxLib 單例化
將 DAQ 串流從「累積全部原始樣本後彙總」改為增量統計模式，大幅降低
長時間錄製的記憶體消耗；同時新增錄製期間即時寫入 CSV/Avro 檔案功能。

DaqmxLib 單例化：
- load() → get_or_load()，使用 OnceLock 保證 nicaiu.dll 全程序僅載入一次
- 消除 Arc<DaqmxLib>，背景執行緒改用 &'static DaqmxLib 參照
- TaskHandle 型別從 u32 改為 usize（匹配 NI-DAQmx 64-bit ABI）

增量統計 (stats.rs)：
- 新增 RunningStats：Welford 式累加器（count/sum/sum_sq/min/max）
- 新增 RunningPower：增量 V*I 平均功率累加器
- DaqSessionResult 不再攜帶 all_samples Vec，改為 channel_stats + power_stats
- stop() 彙總直接使用預計算結果，無需事後遍歷全部樣本

串流檔案輸出：
- 新增 avro_writer.rs 模組，支援 Apache Avro（Deflate 壓縮）格式輸出
  動態建立 schema（timestamp_s + ch_0..ch_N），逐 chunk 寫入
- streaming.rs 新增 DaqFileWriter enum（Csv/Avro），依 output_path 副檔名
  自動選擇格式，每次 poll 後即時 flush
- profiler_daq_start FFI 新增 output_path 參數（UTF-16 指標，可 null）

其他清理：
- reader.rs: 移除未使用的 _reserved 變數，改用 null_mut()
- Cargo.toml: 新增 apache-avro = "0.17" 依賴

+++ src/daq/constants.rs
@@ -47,4 +47,4 @@ pub const DAQmx_Val_WaitInfinitely: f64 = -1.0;
 pub const DAQmx_Val_Acquired_Into_Buffer: i32 = 1;
 
 // Task handle type
-pub type TaskHandle = u32;
+pub type TaskHandle = usize;
